tts-piper/requirements.txt
piper-tts
# onnxruntime-gpu
